refactor: remove non-index re-exports
