+++ libs/mql-interpreter/src/libs/builtins/index.ts
@@ -1,8 +1,6 @@
 import { coreBuiltins, envBuiltins } from "./registry";
 import type { BuiltinFunction } from "./types";
 
-export type { BuiltinFunction } from "./types";
-
 let customEnvBuiltins: Record<string, BuiltinFunction> = {};
 
 export function registerEnvBuiltins(map: Record<string, BuiltinFunction>): void {
@@ -12,5 +10,3 @@ export function registerEnvBuiltins(map: Record<string, BuiltinFunction>): void 
 export function getBuiltin(name: string): BuiltinFunction | undefined {
   return customEnvBuiltins[name] || envBuiltins[name] || coreBuiltins[name];
 }
-
-export { coreBuiltins, envBuiltins };
