--- libs/mql-interpreter/src/runtime/index.ts
@@ -1,9 +1,0 @@
-export { execute, callFunction, instantiate, callMethod } from "./runtime";
-export type {
-  RuntimeState,
-  ExecutionContext,
-  RuntimeFunctionParameter,
-  RuntimeClassField,
-  RuntimeClassMethod,
-  ProgramType,
-} from "./types";
