+++ libs/mql-interpreter/test/runtime/runtime.test.ts
@@ -1,6 +1,6 @@
 import { lex } from "../../src/parser/lexer";
 import { parse } from "../../src/parser/parser";
-import { execute, callFunction, instantiate, callMethod } from "../../src/runtime";
+import { execute, callFunction, instantiate, callMethod } from "../../src/runtime/runtime";
 import { executeStatements } from "../../src/runtime/statements";
 import { describe, it, expect, vi } from "vitest";
